Update copy / design on Medicaid success page
* Make it look more like the SNAP success page
* [Finishes #152423848]
