+++ db/migrate/20171102231334_add_office_location_to_medicaid_applications.rb
@@ -1,0 +1,5 @@
+class AddOfficeLocationToMedicaidApplications < ActiveRecord::Migration[5.1]
+  def change
+    add_column :medicaid_applications, :office_location, :string
+  end
+end
