+++ spec/features/medicaid_application_with_maximum_info_spec.rb
@@ -178,5 +178,11 @@ RSpec.feature "Medicaid app" do
       )
       click_on "I'll do this later"
     end
+
+    on_pages "Application Submitted" do
+      expect(page).to have_content(
+        "Your application has been successfully submitted",
+      )
+    end
   end
 end
